P4ADEV-1275 resolved change request
