+++ src/main/java/it/gov/pagopa/payhub/auth/service/a2a/ClientService.java
@@ -13,5 +13,4 @@ public interface ClientService {
 	String getClientSecret(String organizationIpaCode, String clientId);
 	List<ClientNoSecretDTO> getClients(String organizationIpaCode);
 	Optional<Client> getClientByClientId(String clientId);
-	ClientDTO authorizeCredentials(String clientId, String clientSecret);
 }
